Blueprints: Simplify runPhpWithZipFunctions() setup
This PR removes the duplicate PHP ZIP functions declarations. It only
leaves a single `unzip()` function declaration and a single `zipDir()`
function declaration. Furthermore, they're no longer imported across
modules but live where they're used.

 ## Testing instructions

Confirm the CI checks pass
